# specs/006-workflow-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Integration API
  description: API for integrating textbook content generation workflow with backend
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /api/content/generate:
    post:
      summary: Trigger content generation workflow
      description: |
        Creates a new content generation job and returns job_id for status tracking.
        The workflow executes asynchronously: Content Architect → Template Generator → 
        Technical Writer → Validation → RAG Indexing → Frontend Sync.
      operationId: createContentGenerationJob
      tags:
        - Content Generation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentGenerationRequest'
            example:
              chapters: ["1", "2", "3"]
              lessons: ["1", "2"]
              research_guidance:
                "1.1": "Focus on embodied intelligence definition and sensor-motor integration"
                "1.2": "Cover history from Brooks 1991 to modern robots"
              target_word_count: 800
              priority: "normal"
      responses:
        '202':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreatedResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/content/generate/{job_id}:
    get:
      summary: Get content generation job status
      description: |
        Returns the current status, progress, and results (if completed) for a content generation job.
        Clients should poll this endpoint every 30 seconds for status updates.
      operationId: getContentGenerationJobStatus
      tags:
        - Content Generation
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          description: Job identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/content/generate/{job_id}/cancel:
    post:
      summary: Cancel content generation job
      description: |
        Cancels a pending or in-progress content generation job.
        Completed or failed jobs cannot be cancelled.
      operationId: cancelContentGenerationJob
      tags:
        - Content Generation
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          description: Job identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCancelledResponse'
        '400':
          description: Job cannot be cancelled (already completed/failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ContentGenerationRequest:
      type: object
      required:
        - chapters
      properties:
        chapters:
          type: array
          items:
            type: string
          description: Chapter identifiers (e.g., ["1", "2", "3"] or ["1-3"])
          minItems: 1
        lessons:
          type: array
          items:
            type: string
          description: Lesson identifiers per chapter (e.g., ["1", "2"] or ["1-2"])
        research_guidance:
          type: object
          additionalProperties:
            type: string
          description: Research guidance per lesson (lesson_id → guidance text)
        target_word_count:
          type: integer
          minimum: 500
          maximum: 1000
          default: 800
          description: Target words per lesson
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
          description: Job priority

    JobCreatedResponse:
      type: object
      required:
        - job_id
        - status
        - message
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique job identifier
        status:
          type: string
          enum: [pending]
          description: Initial job status
        message:
          type: string
          description: Success message
        estimated_completion_time:
          type: integer
          description: Estimated completion time in seconds

    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
        - progress
        - created_at
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, in_progress, validating, indexing, syncing, completed, failed, cancelled]
        progress:
          type: number
          format: float
          minimum: 0.0
          maximum: 100.0
        current_step:
          type: string
          description: Current workflow step description
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        request:
          $ref: '#/components/schemas/ContentGenerationRequest'
        results:
          $ref: '#/components/schemas/ContentGenerationResults'
        error:
          $ref: '#/components/schemas/ErrorDetails'
        component_statuses:
          type: object
          additionalProperties:
            type: string
            enum: [pending, in_progress, completed, failed]
          description: Status of each workflow component
        validation_results:
          $ref: '#/components/schemas/ValidationResult'
        indexing_status:
          type: string
          enum: [pending, in_progress, completed, failed]
        sync_status:
          type: string
          enum: [pending, in_progress, completed, failed]

    ContentGenerationResults:
      type: object
      required:
        - total_lessons
        - successful_lessons
        - failed_lessons
      properties:
        lessons_generated:
          type: array
          items:
            $ref: '#/components/schemas/LessonResult'
        total_lessons:
          type: integer
        successful_lessons:
          type: integer
        failed_lessons:
          type: integer
        content_locations:
          type: array
          items:
            type: string
          description: File paths to generated content
        generation_time:
          type: number
          format: float
          description: Total generation time in seconds
        warnings:
          type: array
          items:
            type: string

    LessonResult:
      type: object
      required:
        - lesson_id
        - status
      properties:
        lesson_id:
          type: string
        status:
          type: string
          enum: [completed, failed, skipped]
        file_path:
          type: string
        word_count:
          type: integer
        sources_used:
          type: integer
        error:
          type: string
        edge_cases_handled:
          type: array
          items:
            type: string

    ValidationResult:
      type: object
      required:
        - overall_status
      properties:
        overall_status:
          type: string
          enum: [passed, failed, warning]
        word_count_validation:
          $ref: '#/components/schemas/WordCountValidation'
        source_validation:
          $ref: '#/components/schemas/SourceValidation'
        quality_checks:
          type: array
          items:
            $ref: '#/components/schemas/QualityCheck'
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string

    WordCountValidation:
      type: object
      required:
        - status
        - target_word_count
        - actual_word_count
        - variance
      properties:
        status:
          type: string
          enum: [passed, failed, warning]
        target_word_count:
          type: integer
        actual_word_count:
          type: integer
        variance:
          type: number
          format: float
        message:
          type: string

    SourceValidation:
      type: object
      required:
        - status
        - sources_checked
        - authoritative_sources
        - non_authoritative_sources
        - min_required
      properties:
        status:
          type: string
          enum: [passed, failed, warning]
        sources_checked:
          type: integer
        authoritative_sources:
          type: integer
        non_authoritative_sources:
          type: integer
        min_required:
          type: integer
        message:
          type: string

    QualityCheck:
      type: object
      required:
        - check_name
        - status
      properties:
        check_name:
          type: string
        status:
          type: string
          enum: [passed, failed, warning]
        message:
          type: string

    ErrorDetails:
      type: object
      required:
        - error_type
        - message
        - occurred_at
      properties:
        error_type:
          type: string
          enum: [insufficient_sources, component_failure, validation_error, workflow_error, timeout_error]
        component:
          type: string
        message:
          type: string
        recovery_suggestions:
          type: array
          items:
            type: string
        retryable:
          type: boolean
        occurred_at:
          type: string
          format: date-time

    JobCancelledResponse:
      type: object
      required:
        - job_id
        - status
        - message
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [cancelled]
        message:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true